requirements.txt
streamlit==1.29.0
Pillow==10.1.0
torch==2.2.0
torchvision==0.17.0
ftfy==6.1.1
regex==2023.10.3
tqdm==4.66.1
numpy==1.26.0